refactor(services): extract service config tests into dedicated file
